fix：修复发红包不能用微信支付宝问题

+++ service-share/ad-srv/src/test/java/com/lawu/eshop/ad/srv/service/impl/UserRedPacketServiceImplTest.java
@@ -56,7 +56,7 @@ public class UserRedPacketServiceImplTest {
 		param.setTotalMoney(new BigDecimal(100));
 		param.setUserAccount("15000000000");
 		param.setUserNum("M000001");
-		Integer i = userRedPacketService.addUserRedPacket(param);
+		Long i = userRedPacketService.addUserRedPacket(param);
 		UserRedPacketSelectParam query = new UserRedPacketSelectParam();
 		query.setCurrentPage(1);
 		query.setPageSize(10);
